requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
python-dotenv>=1.0.0
django-filter>=23.1

# Image Processing
Pillow>=10.0.0

# Date/Time
python-dateutil>=2.8.2

# Multi-tenancy
django-tenants>=3.5.0

# Database
dj-database-url>=1.2.0
psycopg2-binary>=2.9.7

# Authentication
django-allauth>=0.54.0
django-rest-auth>=0.9.5

# API Documentation
drf-yasg>=1.21.5

# Static Files
whitenoise>=6.5.0

# Health Checks
django-health-check>=3.17.0

# WSGI Server
gunicorn>=20.1.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# System Monitoring
psutil>=5.9.0

# PDF Generation
xhtml2pdf>=0.2.17
reportlab>=4.0.0

# AWS Integration
boto3>=1.28.0
django-storages>=1.14.0

# Email
django-ses>=3.5.0

# Security
django-ratelimit>=4.1.0
django-defender>=0.9.0

# Monitoring & Logging
sentry-sdk>=1.40.0

# Task Queue (Optional - for background jobs)
celery>=5.3.0
redis>=5.0.0

# Environment Management
python-decouple>=3.8
